+++ src/Services/Store/Core/Store.Application/Features/Products/Commands/DeleteProductCommand/DeleteProductCommand.cs
@@ -7,6 +7,8 @@ public class DeleteProductCommandHandler(IApplicationDbContext dbContext) : IReq
     public async Task<bool> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
         var product = await dbContext.Products.FirstOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
+        if (product is null)
+            throw new NotFoundException($"Product not found with ID: {request.Id}");
         dbContext.Products.Remove(product);
 
         return await dbContext.SaveChangesAsync(cancellationToken) > 0;
